# Rewrite delta to use only a single goroutine